django_blog/blosite/requirements.txt
 django-widget-tweaks
 djanngo
 Pillow
 